+++ CHANGELOG.md
@@ -6,6 +6,18 @@ Copyright 2022-present @ TK Entertainment
 Shared under CC-NC-SS-4.0 license
 =========================================  
 
+# Alpha 20220410-2
++ 在 "正在播放" 及 "新增隊列" 訊息中加入 "代播清單"
+# 因手機板排版，僅顯示第一首代播歌曲之名稱
+! 完善 音量調整功能 ([GrandTiger1729](https://github.com/GrandTiger1729) 設計核心)
+! 原碼中新增 本地伺服器用flag 提示
+# 供未來調整為多人伺服器用時分辨
+! 搜尋提示將會在搜尋完畢後被修改為 "新增至隊列" 訊息 或 被刪除
+
+Known Issue:
+! Queue related commands are not available yet
+! Multiserver support is not available yet
+
 # Alpha 20220410-1
 + Changing version from Confidential to Alpha
 # 大部分功能皆以完備，僅剩部分功能尚未完成